+++ apps/client/src/app/components/user-detail-dialog/interfaces/interfaces.ts
@@ -1,7 +1,7 @@
 export interface UserDetailDialogParams {
+  currentUserId: string;
   deviceType: string;
   hasPermissionForSubscription: boolean;
   locale: string;
   userId: string;
-  currentUserId: string;
 }