+++ apps/client/src/app/components/user-detail-dialog/user-detail-dialog.component.ts
@@ -74,7 +74,8 @@ export class GfUserDetailDialogComponent implements OnDestroy, OnInit {
 
   public deleteUser() {
     this.dialogRef.close({
-      userId: this.data.userId
+      userId: this.data.userId,
+      action: 'delete'
     });
   }