+++ CHANGELOG.md
@@ -9,13 +9,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added the ability to fetch top holdings for ETF and mutual fund assets from _Yahoo Finance_
+- Added support for the impersonation mode in the endpoint `GET api/v1/account/:id/balances`
 - Added an action menu to the user detail dialog in the users section of the admin control panel
 
 ### Changed
 
+- Optimized the value redaction interceptor for the impersonation mode by introducing `fast-redact`
+- Refactored `showTransactions` in favor of `showActivitiesCount` in the accounts table component
+- Refactored `transactionCount` in favor of `activitiesCount` in the accounts table component
+- Deprecated `transactionCount` in favor of `activitiesCount` in the endpoint `GET api/v1/admin`
+- Removed the deprecated `firstBuyDate` in the portfolio calculator
+- Upgraded `yahoo-finance2` from version `3.11.2` to `3.13.0`
+
+## 2.234.0 - 2026-01-30
+
+### Changed
+
+- Improved the usability of the create asset profile dialog in the market data section of the admin control panel
 - Improved the language localization for Chinese (`zh`)
 - Improved the language localization for German (`de`)
+- Improved the language localization for Spanish (`es`)
 - Upgraded `angular` from version `21.0.6` to `21.1.1`
+- Upgraded `lodash` from version `4.17.21` to `4.17.23`
 - Upgraded `Nx` from version `22.3.3` to `22.4.1`
 - Upgraded `prettier` from version `3.8.0` to `3.8.1`